*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@934 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tast/submission/SubmissionBean.java
@@ -58,7 +58,7 @@ public class SubmissionBean
 	
 	private RowGroup[] rowGroups;
 
-	private Long lookupVoyageId = null;
+	private Integer lookupVoyageId = null;
 	private boolean lookupPerformed = false;
 	private SelectedVoyageInfo lookedUpVoyage = null;
 
@@ -119,7 +119,7 @@ public class SubmissionBean
 			if (selectedVoyageForEdit != null)
 			{
 				lookedUpVoyage = selectedVoyageForEdit;
-				lookupVoyageId = new Long(selectedVoyageForEdit.getVoyageId());
+				lookupVoyageId = new Integer(selectedVoyageForEdit.getVoyageId());
 			}
 		}
 		return "edit";
@@ -600,12 +600,12 @@ public class SubmissionBean
 		return SubmissionDictionaries.fieldTypes;
 	}
 
-	public Long getLookupVoyageId()
+	public Integer getLookupVoyageId()
 	{
 		return lookupVoyageId;
 	}
 
-	public void setLookupVoyageId(Long voyageId)
+	public void setLookupVoyageId(Integer voyageId)
 	{
 		this.lookupVoyageId = voyageId;
 	}
